# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.runRemindersRequest:
    properties:
      min_pending_age_sec:
        type: integer
      reminder_type:
        type: string
      window_truncation_sec:
        type: integer
    type: object
  handlers.startWorkflowRequest:
    properties:
      comment:
        type: string
      definition_code:
        type: string
      document_id:
        type: integer
      document_type:
        type: string
    required:
    - definition_code
    - document_id
    - document_type
    type: object
...
  handlers.swaggerEnvelope:
    type: object
  handlers.swaggerMessageResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  handlers.workflowActionRequest:
    properties:
      comment:
        type: string
      idempotency_key:
        type: string
    required:
    - idempotency_key
    type: object
...
  workflow.Action:
    enum:
    - submit
    - approve
    - reject
    - resubmit
    type: string
    x-enum-varnames:
    - ActionSubmit
    - ActionApprove
    - ActionReject
    - ActionResubmit
  workflow.AuditEntry:
    properties:
      action:
        $ref: '#/definitions/workflow.Action'
      actor_user_id:
        type: integer
      comment:
        type: string
      created_at:
        type: string
      from_status:
        $ref: '#/definitions/workflow.InstanceStatus'
      from_step_order:
        type: integer
      id:
        type: integer
      idempotency_key:
        type: string
      to_status:
        $ref: '#/definitions/workflow.InstanceStatus'
      to_step_order:
        type: integer
      workflow_instance_id:
        type: integer
    type: object
  workflow.Definition:
    properties:
      code:
        type: string
      id:
        format: int64
        type: integer
      name:
        type: string
      nodes:
        items:
          $ref: '#/definitions/workflow.Node'
        type: array
      processClass:
        type: string
      transitions:
        items:
          $ref: '#/definitions/workflow.Transition'
        type: array
    type: object
  workflow.Instance:
    properties:
      completed_at:
        type: string
      current_cycle:
        type: integer
      current_node_id:
        type: integer
      current_step_order:
        type: integer
      document_id:
        type: integer
      document_type:
        type: string
      id:
        type: integer
      status:
        $ref: '#/definitions/workflow.InstanceStatus'
      submitted_at:
        type: string
      submitted_by:
        type: integer
      version:
        type: integer
      workflow_definition_id:
        type: integer
    type: object
  workflow.InstanceStatus:
    enum:
    - pending
    - approved
    - rejected
    type: string
    x-enum-varnames:
    - InstanceStatusPending
    - InstanceStatusApproved
    - InstanceStatusRejected
  workflow.Node:
    properties:
      canSubmitToManager:
        type: boolean
      code:
        type: string
      functionID:
        format: int64
        type: integer
      id:
        format: int64
        type: integer
      name:
        type: string
      printsAfterConfirm:
        type: boolean
      processClass:
        type: string
      roleID:
        format: int64
        type: integer
      stepOrder:
        type: integer
      validatesAfterConfirm:
        type: boolean
      workflowDefinitionID:
        format: int64
        type: integer
    type: object
  workflow.ReminderAuditRecord:
    properties:
      created_at:
        type: string
      id:
        type: integer
      outcome:
        $ref: '#/definitions/workflow.ReminderOutcome'
      reason_code:
        type: string
      reminder_key:
        type: string
      reminder_type:
        type: string
      reminder_window_start:
        type: string
      workflow_instance_id:
        type: integer
    type: object
  workflow.ReminderOutcome:
    enum:
    - emitted
    - skipped
    type: string
    x-enum-varnames:
    - ReminderOutcomeEmitted
    - ReminderOutcomeSkipped
  workflow.Transition:
    properties:
      action:
        $ref: '#/definitions/workflow.Action'
      fromNodeID:
        format: int64
        type: integer
      id:
        format: int64
        type: integer
      toNodeID:
        format: int64
        type: integer
      workflowDefinitionID:
        format: int64
        type: integer
    type: object
host: localhost:5180
info:
  contact: {}
  description: Backend API for the MI property management system
  title: MI Management Information System API
  version: "1.0"
paths:
...
  /workflow/definitions:
    get:
      consumes:
      - application/json
      description: Returns workflow definitions available for approval flows.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.swaggerEnvelope'
            - properties:
                definitions:
                  items:
                    $ref: '#/definitions/workflow.Definition'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
      security:
      - BearerAuth: []
      summary: List workflow definitions
      tags:
      - Workflow
  /workflow/instances:
    get:
      consumes:
      - application/json
      description: Returns workflow instances filtered by status, document type, and
        document ID.
      parameters:
      - description: Workflow status
        in: query
        name: status
        type: string
      - description: Document type
        in: query
        name: document_type
        type: string
      - description: Document ID
        in: query
        name: document_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.swaggerEnvelope'
            - properties:
                instances:
                  items:
                    $ref: '#/definitions/workflow.Instance'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
      security:
      - BearerAuth: []
      summary: List workflow instances
      tags:
      - Workflow
    post:
      consumes:
      - application/json
      description: Starts a workflow instance for a document and assigns the first
        approval step.
      parameters:
      - description: Workflow start request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.startWorkflowRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handlers.swaggerEnvelope'
            - properties:
                instance:
                  $ref: '#/definitions/workflow.Instance'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
      security:
      - BearerAuth: []
      summary: Start workflow instance
      tags:
      - Workflow
  /workflow/instances/{id}:
    get:
      consumes:
      - application/json
      description: Returns a workflow instance by ID.
      parameters:
      - description: Workflow instance ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.swaggerEnvelope'
            - properties:
                instance:
                  $ref: '#/definitions/workflow.Instance'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
      security:
      - BearerAuth: []
      summary: Get workflow instance
      tags:
      - Workflow
  /workflow/instances/{id}/approve:
    post:
      consumes:
      - application/json
      description: Approves the current step of a workflow instance.
      parameters:
      - description: Workflow instance ID
        in: path
        name: id
        required: true
        type: integer
      - description: Workflow action request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.workflowActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.swaggerEnvelope'
            - properties:
                instance:
                  $ref: '#/definitions/workflow.Instance'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
      security:
      - BearerAuth: []
      summary: Approve workflow instance
      tags:
      - Workflow
  /workflow/instances/{id}/audit:
    get:
      consumes:
      - application/json
      description: Returns workflow audit entries for a workflow instance.
      parameters:
      - description: Workflow instance ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.swaggerEnvelope'
            - properties:
                history:
                  items:
                    $ref: '#/definitions/workflow.AuditEntry'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
      security:
      - BearerAuth: []
      summary: Get workflow audit history
      tags:
      - Workflow
  /workflow/instances/{id}/reject:
    post:
      consumes:
      - application/json
      description: Rejects the current step of a workflow instance.
      parameters:
      - description: Workflow instance ID
        in: path
        name: id
        required: true
        type: integer
      - description: Workflow action request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.workflowActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.swaggerEnvelope'
            - properties:
                instance:
                  $ref: '#/definitions/workflow.Instance'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
      security:
      - BearerAuth: []
      summary: Reject workflow instance
      tags:
      - Workflow
  /workflow/instances/{id}/reminders:
    get:
      consumes:
      - application/json
      description: Returns reminder audit entries for a workflow instance.
      parameters:
      - description: Workflow instance ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.swaggerEnvelope'
            - properties:
                reminders:
                  items:
                    $ref: '#/definitions/workflow.ReminderAuditRecord'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
      security:
      - BearerAuth: []
      summary: Get workflow reminder history
      tags:
      - Workflow
  /workflow/instances/{id}/resubmit:
    post:
      consumes:
      - application/json
      description: Resubmits a rejected workflow instance for approval.
      parameters:
      - description: Workflow instance ID
        in: path
        name: id
        required: true
        type: integer
      - description: Workflow action request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.workflowActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.swaggerEnvelope'
            - properties:
                instance:
                  $ref: '#/definitions/workflow.Instance'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
      security:
      - BearerAuth: []
      summary: Resubmit workflow instance
      tags:
      - Workflow
  /workflow/reminders/run:
    post:
      consumes:
      - application/json
      description: Evaluates pending workflow reminders and records any reminder emissions.
      parameters:
      - description: Reminder execution request
        in: body
        name: request
        schema:
          $ref: '#/definitions/handlers.runRemindersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.swaggerEnvelope'
            - properties:
                reminders:
                  items:
                    $ref: '#/definitions/workflow.ReminderAuditRecord'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
      security:
      - BearerAuth: []
      summary: Run workflow reminders
      tags:
      - Workflow
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"